Set TODO reminder comments
As I tend to have to switch back and forth between my leisure work and
serious work (working at a startup gets busy), I wrote better notes to
help me remember where I left off. This is a good practice to adopt
when working with others on the same project as well.